requirement.txt
gunicorn==20.1.0
Django==3.2.6
djangorestframework==3.13.1
psycopg2==2.9.1
whitenoise==5.3.0
django-heroku==0.3.1